Refactor phase execution into SRP helpers
