+++ crates/xchecker-engine/src/orchestrator/phase_exec.rs
@@ -10,15 +10,25 @@ use anyhow::{Context, Result};
 use crate::error::{PhaseError, XCheckerError};
 use crate::exit_codes;
 use crate::fixup::{FixupMode, FixupPhase};
-use crate::hooks::{HookContext, HookExecutor, HookType, execute_and_process_hook};
-use crate::packet::PacketBuilder;
 use crate::phase::{Phase, PhaseContext};
 use crate::phases::{DesignPhase, RequirementsPhase, ReviewPhase, TasksPhase};
 use crate::status::artifact::{Artifact, ArtifactType};
-use crate::types::{ErrorKind, FileType, LlmInfo, PacketEvidence, PhaseId, PipelineInfo};
+use crate::types::{ErrorKind, PacketEvidence, PhaseId, PipelineInfo};
 
-use super::llm::{ClaudeExecutionMetadata, LlmInvocationError};
+use super::llm::ClaudeExecutionMetadata;
 use super::{OrchestratorConfig, PhaseOrchestrator, PhaseTimeout};
+
+#[path = "phase_exec/artifacts.rs"]
+mod artifacts;
+#[path = "phase_exec/hooks.rs"]
+mod hooks;
+#[path = "phase_exec/llm.rs"]
+mod llm;
+#[path = "phase_exec/packet.rs"]
+mod packet;
+
+use hooks::PrePhaseHookResult;
+use llm::PhaseLlmOutcome;
 
 /// Result of executing a phase through the orchestrator.
 ///
@@ -502,26 +512,7 @@ impl PhaseOrchestrator {
         let prompt = phase.prompt(&phase_context);
 
         // Step 2: Build packet (FR-ORC-003)
-        let packet = phase.make_packet(&phase_context).map_err(|e| {
-            XCheckerError::Phase(PhaseError::PacketCreationFailed {
-                phase: phase_id.as_str().to_string(),
-                reason: e.to_string(),
-            })
-        })?;
-
-        // Log packet hash and budget usage for visibility
-        let budget = packet.budget_usage();
-        tracing::info!(
-            target: "xchecker::packet",
-            spec_id = %self.spec_id(),
-            phase = %phase_id.as_str(),
-            packet_hash = %packet.hash(),
-            bytes_used = budget.bytes_used,
-            bytes_limit = budget.max_bytes,
-            lines_used = budget.lines_used,
-            lines_limit = budget.max_lines,
-            "Built packet for phase"
-        );
+        let packet = self.build_phase_packet(phase, &phase_context, phase_id)?;
 
         let packet_evidence = packet.evidence.clone();
 
@@ -537,37 +528,8 @@ impl PhaseOrchestrator {
             .into());
         }
 
-        // Store packet for debugging/preview
-        let _packet_preview_path = self
-            .artifact_manager()
-            .store_context_file(&format!("{}-packet", phase_id.as_str()), &packet.content)?;
-
-        // Step 4: Write full debug packet if --debug-packet flag is set (FR-PKT-006, FR-PKT-007)
-        // Only write after secret scan passes; file is excluded from receipts
-        let debug_packet_enabled = config
-            .config
-            .get("debug_packet")
-            .is_some_and(|s| s == "true");
-
-        if debug_packet_enabled {
-            // Get context directory from artifact manager
-            let context_dir = self.artifact_manager().context_path();
-
-            // Create a temporary PacketBuilder just to call write_debug_packet
-            let temp_builder = PacketBuilder::new().map_err(|e| {
-                XCheckerError::Phase(PhaseError::PacketCreationFailed {
-                    phase: phase_id.as_str().to_string(),
-                    reason: format!("Failed to create PacketBuilder for debug packet: {e}"),
-                })
-            })?;
-
-            if let Err(e) =
-                temp_builder.write_debug_packet(&packet.content, phase_id.as_str(), &context_dir)
-            {
-                // Log warning but don't fail the operation (debug packet is optional)
-                eprintln!("Warning: Failed to write debug packet: {e}");
-            }
-        }
+        // Store packet for debugging/preview after secret scan passes.
+        self.persist_packet_context(phase_id, &packet.content, config)?;
 
         // Step 5: Execute LLM (or simulate in dry-run mode)
         let (claude_response, claude_exit_code, claude_metadata, llm_result, llm_fallback_warning) =
@@ -614,24 +576,8 @@ impl PhaseOrchestrator {
             }
         };
 
-        // Step 7: Write partial artifacts to .partial/ subdirectory (FR-ORC-004)
-        for artifact in &phase_result.artifacts {
-            // Store to .partial/ staging directory first
-            let _partial_result = self
-                .artifact_manager()
-                .store_partial_staged_artifact(artifact)
-                .with_context(|| format!("Failed to store partial artifact: {}", artifact.name))?;
-        }
-
-        // Step 8: Promote to final (atomic rename) (FR-ORC-004)
-        for artifact in &phase_result.artifacts {
-            let _final_path = self
-                .artifact_manager()
-                .promote_staged_to_final(&artifact.name)
-                .with_context(|| {
-                    format!("Failed to promote artifact to final: {}", artifact.name)
-                })?;
-        }
+        // Steps 7-8: Stage artifacts, then promote them to final paths (FR-ORC-004).
+        self.store_core_phase_artifacts(&phase_result.artifacts)?;
 
         // Return intermediate values for receipt generation
         // Note: Callers (workflow.rs) re-store artifacts and re-compute hashes,
@@ -671,175 +617,19 @@ impl PhaseOrchestrator {
         // Check dependencies (Requirements phase has no deps)
         self.check_phase_dependencies(phase)?;
 
-        // Execute pre-phase hook if configured
-        // Hooks run from invocation CWD so relative paths like ./scripts/... work
-        let mut hook_warnings: Vec<String> = Vec::new();
-        if let Some(ref hooks_config) = config.hooks
-            && let Some(hook_config) = hooks_config.get_pre_phase_hook(phase_id)
+        let hook_warnings = match self
+            .run_pre_phase_hook(phase_id, config, pipeline_info.clone())
+            .await?
         {
-            let executor = HookExecutor::new(
-                std::env::current_dir().unwrap_or_else(|_| std::path::PathBuf::from(".")),
-            );
-            let context = HookContext::new(self.spec_id(), phase_id, HookType::PrePhase);
-
-            match execute_and_process_hook(
-                &executor,
-                hook_config,
-                &context,
-                HookType::PrePhase,
-                phase_id,
-            )
-            .await
-            {
-                Ok(outcome) => {
-                    if let Some(warning) = outcome.warning() {
-                        hook_warnings.push(warning.to_warning_string());
-                    }
-                    if !outcome.should_continue() {
-                        // Pre-hook failed with on_fail=fail - abort phase but still create receipt
-                        let error_reason = format!(
-                            "Pre-phase hook failed: {}",
-                            outcome.error().map(|e| e.to_string()).unwrap_or_default()
-                        );
-
-                        // Create failure receipt for hook failure (audit trail requirement)
-                        let packet_evidence = PacketEvidence {
-                            files: vec![],
-                            max_bytes: 65536,
-                            max_lines: 1200,
-                        };
-                        let mut flags = HashMap::new();
-                        flags.insert("phase".to_string(), phase_id.as_str().to_string());
-                        flags.insert("hook_failure".to_string(), "pre_phase".to_string());
-
-                        // Use config values for truthful failure receipts (no hard-coded metadata)
-                        let configured_model =
-                            config.config.get("model").map_or("unknown", |s| s.as_str());
-                        let configured_runner = config
-                            .config
-                            .get("runner_mode")
-                            .map_or("unknown", |s| s.as_str());
-
-                        let receipt = self.receipt_manager().create_receipt_with_redactor(
-                            config.redactor.as_ref(),
-                            self.spec_id(),
-                            phase_id,
-                            exit_codes::codes::CLAUDE_FAILURE,
-                            vec![], // No outputs
-                            env!("CARGO_PKG_VERSION"),
-                            "unknown", // Claude hasn't run yet, so version is unknown
-                            configured_model,
-                            None,
-                            flags,
-                            packet_evidence,
-                            None,
-                            None,
-                            hook_warnings.clone(),
-                            None,
-                            configured_runner,
-                            None,
-                            Some(ErrorKind::ClaudeFailure),
-                            Some(error_reason.clone()),
-                            None,
-                            pipeline_info.clone(),
-                        );
-
-                        let receipt_path = self.receipt_manager().write_receipt(&receipt)?;
-
-                        return Ok(ExecutionResult {
-                            phase: phase_id,
-                            success: false,
-                            exit_code: exit_codes::codes::CLAUDE_FAILURE,
-                            artifact_paths: vec![],
-                            receipt_path: Some(receipt_path.into_std_path_buf()),
-                            error: Some(error_reason),
-                        });
-                    }
-                }
-                Err(e) => {
-                    // Hook execution error - treat as failure but still create receipt
-                    let error_reason = format!("Pre-phase hook error: {}", e);
-
-                    // Create failure receipt for hook error (audit trail requirement)
-                    let packet_evidence = PacketEvidence {
-                        files: vec![],
-                        max_bytes: 65536,
-                        max_lines: 1200,
-                    };
-                    let mut flags = HashMap::new();
-                    flags.insert("phase".to_string(), phase_id.as_str().to_string());
-                    flags.insert("hook_error".to_string(), "pre_phase".to_string());
-
-                    // Use config values for truthful failure receipts (no hard-coded metadata)
-                    let configured_model =
-                        config.config.get("model").map_or("unknown", |s| s.as_str());
-                    let configured_runner = config
-                        .config
-                        .get("runner_mode")
-                        .map_or("unknown", |s| s.as_str());
-
-                    let receipt = self.receipt_manager().create_receipt_with_redactor(
-                        config.redactor.as_ref(),
-                        self.spec_id(),
-                        phase_id,
-                        exit_codes::codes::CLAUDE_FAILURE,
-                        vec![], // No outputs
-                        env!("CARGO_PKG_VERSION"),
-                        "unknown", // Claude hasn't run yet, so version is unknown
-                        configured_model,
-                        None,
-                        flags,
-                        packet_evidence,
-                        None,
-                        None,
-                        vec![format!("hook_error:pre_phase:{}", e)],
-                        None,
-                        configured_runner,
-                        None,
-                        Some(ErrorKind::ClaudeFailure),
-                        Some(error_reason.clone()),
-                        None,
-                        pipeline_info.clone(),
-                    );
-
-                    let receipt_path = self.receipt_manager().write_receipt(&receipt)?;
-
-                    return Ok(ExecutionResult {
-                        phase: phase_id,
-                        success: false,
-                        exit_code: exit_codes::codes::CLAUDE_FAILURE,
-                        artifact_paths: vec![],
-                        receipt_path: Some(receipt_path.into_std_path_buf()),
-                        error: Some(error_reason),
-                    });
-                }
-            }
-        }
+            PrePhaseHookResult::Continue { warnings } => warnings,
+            PrePhaseHookResult::Abort(result) => return Ok(result),
+        };
 
         // Generate prompt
         let prompt = phase.prompt(&phase_context);
 
         // Step 3: Build packet (FR-ORC-003)
-        let packet = phase.make_packet(&phase_context).map_err(|e| {
-            XCheckerError::Phase(PhaseError::PacketCreationFailed {
-                phase: phase_id.as_str().to_string(),
-                reason: e.to_string(),
-            })
-        })?;
-
-        // Log packet hash and budget usage for visibility
-        let budget = packet.budget_usage();
-        tracing::info!(
-            target: "xchecker::packet",
-            spec_id = %self.spec_id(),
-            phase = %phase_id.as_str(),
-            packet_hash = %packet.hash(),
-            bytes_used = budget.bytes_used,
-            bytes_limit = budget.max_bytes,
-            lines_used = budget.lines_used,
-            lines_limit = budget.max_lines,
-            "Built packet for phase"
-        );
+        let packet = self.build_phase_packet(phase, &phase_context, phase_id)?;
 
         // Step 4: Scan for secrets (FR-ORC-003, FR-SEC)
         let redactor = config.redactor.as_ref();
@@ -896,257 +686,36 @@ impl PhaseOrchestrator {
             });
         }
 
-        // Store packet for debugging/preview
-        let _packet_preview_path = self
-            .artifact_manager()
-            .store_context_file(&format!("{}-packet", phase_id.as_str()), &packet.content)?;
+        // Store packet for debugging/preview after secret scan passes.
+        self.persist_packet_context(phase_id, &packet.content, config)?;
 
-        // Write full debug packet if --debug-packet flag is set (FR-PKT-006, FR-PKT-007)
-        // Only write after secret scan passes; file is excluded from receipts
-        let debug_packet_enabled = config
-            .config
-            .get("debug_packet")
-            .is_some_and(|s| s == "true");
-
-        if debug_packet_enabled {
-            // Get context directory from artifact manager
-            let context_dir = self.artifact_manager().context_path();
-
-            // Create a temporary PacketBuilder just to call write_debug_packet
-            // (This is a bit awkward but maintains the existing API)
-            let temp_builder = PacketBuilder::new().map_err(|e| {
-                XCheckerError::Phase(PhaseError::PacketCreationFailed {
-                    phase: phase_id.as_str().to_string(),
-                    reason: format!("Failed to create PacketBuilder for debug packet: {e}"),
-                })
-            })?;
-
-            if let Err(e) =
-                temp_builder.write_debug_packet(&packet.content, phase_id.as_str(), &context_dir)
-            {
-                // Log warning but don't fail the operation (debug packet is optional)
-                eprintln!("Warning: Failed to write debug packet: {e}");
-            }
-        }
-
-        // Execute LLM (or simulate in dry-run mode)
-        let mut llm_fallback_warning: Option<String> = None;
-        let (claude_response, claude_exit_code, claude_metadata, llm_result) = if config.dry_run {
-            let simulated_llm = self.simulate_llm_result(phase_id);
-            let simulated_metadata = super::llm::ClaudeExecutionMetadata {
-                model_alias: None,
-                model_full_name: "haiku".to_string(),
-                claude_cli_version: "0.8.1".to_string(),
-                fallback_used: false,
-                runner: "simulated".to_string(),
-                runner_distro: None,
-                stderr_tail: None,
-            };
-            (
-                self.simulate_claude_response(phase_id, &prompt),
-                0,
-                Some(simulated_metadata),
-                Some(simulated_llm),
-            )
-        } else {
-            // Use new LLM backend abstraction (V11: Claude CLI only)
+        // Execute LLM (or simulate in dry-run mode) and convert invocation errors to receipts.
+        let (claude_response, claude_exit_code, claude_metadata, llm_result, llm_fallback_warning) =
             match self
-                .run_llm_invocation(&prompt, &packet.content, phase_id, config)
-                .await
+                .execute_llm_for_phase(phase_id, &prompt, &packet, config, pipeline_info.clone())
+                .await?
             {
-                Ok((response, exit_code, metadata, result, fallback_warning)) => {
-                    llm_fallback_warning = fallback_warning;
-                    (response, exit_code, metadata, result)
-                }
-                Err(e) => {
-                    let (xchecker_err, fallback_warning) =
-                        if let Some(invocation_err) = e.downcast_ref::<LlmInvocationError>() {
-                            (
-                                invocation_err.error(),
-                                invocation_err.fallback_warning().map(|s| s.to_string()),
-                            )
-                        } else if let Some(xchecker_err) = e.downcast_ref::<XCheckerError>() {
-                            (xchecker_err, None)
-                        } else {
-                            return Err(e);
-                        };
-
-                    llm_fallback_warning = fallback_warning;
-
-                    // Check if this is a budget exhaustion error by downcasting
-                    if let XCheckerError::Llm(llm_err) = xchecker_err {
-                        if matches!(llm_err, crate::llm::LlmError::BudgetExceeded { .. }) {
-                            // Handle budget exhaustion specially - create receipt with budget_exhausted flag
-                            let packet_evidence = packet.evidence.clone();
-                            let mut flags = HashMap::new();
-                            flags.insert("phase".to_string(), phase_id.as_str().to_string());
-
-                            // Use config values for truthful failure receipts (no hard-coded metadata)
-                            let configured_model =
-                                config.config.get("model").map_or("unknown", |s| s.as_str());
-                            let configured_runner = config
-                                .config
-                                .get("runner_mode")
-                                .map_or("unknown", |s| s.as_str());
-
-                            let mut warnings = vec![format!("LLM budget exhausted: {}", llm_err)];
-                            if let Some(ref warning) = llm_fallback_warning {
-                                warnings.push(warning.clone());
-                            }
-
-                            let mut receipt = self.receipt_manager().create_receipt_with_redactor(
-                                config.redactor.as_ref(),
-                                self.spec_id(),
-                                phase_id,
-                                exit_codes::codes::CLAUDE_FAILURE, // Exit code 70
-                                vec![],                            // No successful outputs
-                                env!("CARGO_PKG_VERSION"),
-                                "unknown", // Claude hasn't completed, so version is unknown
-                                configured_model,
-                                None, // No model alias
-                                flags,
-                                packet_evidence,
-                                None, // No stderr_tail
-                                None, // No stderr_redacted
-                                warnings,
-                                None, // No fallback
-                                configured_runner,
-                                None, // No runner distro
-                                Some(ErrorKind::ClaudeFailure),
-                                Some(llm_err.to_string()),
-                                None, // No diff_context
-                                pipeline_info.clone(),
-                            );
-
-                            // Attach LlmInfo with budget_exhausted flag
-                            receipt.llm = Some(LlmInfo::for_budget_exhaustion());
-
-                            let receipt_path = self.receipt_manager().write_receipt(&receipt)?;
-
-                            return Ok(ExecutionResult {
-                                phase: phase_id,
-                                success: false,
-                                exit_code: exit_codes::codes::CLAUDE_FAILURE,
-                                artifact_paths: vec![],
-                                receipt_path: Some(receipt_path.into_std_path_buf()),
-                                error: Some(llm_err.to_string()),
-                            });
-                        }
-
-                        let packet_evidence = packet.evidence.clone();
-                        let mut flags = HashMap::new();
-                        flags.insert("phase".to_string(), phase_id.as_str().to_string());
-
-                        // Use config values for truthful failure receipts (no hard-coded metadata)
-                        let configured_model =
-                            config.config.get("model").map_or("unknown", |s| s.as_str());
-                        let configured_runner = config
-                            .config
-                            .get("runner_mode")
-                            .map_or("unknown", |s| s.as_str());
-
-                        let (exit_code, error_kind) =
-                            exit_codes::error_to_exit_code_and_kind(xchecker_err);
-
-                        let invocation =
-                            self.build_llm_invocation(phase_id, &prompt, &packet.content, config);
-                        let provider = self
-                            .config_from_orchestrator_config(config)
-                            .llm
-                            .provider
-                            .unwrap_or_else(|| "claude-cli".to_string());
-
-                        let mut llm_info = LlmInfo {
-                            provider: Some(provider),
-                            model_used: if invocation.model.is_empty() {
-                                None
-                            } else {
-                                Some(invocation.model.clone())
-                            },
-                            tokens_input: None,
-                            tokens_output: None,
-                            timed_out: None,
-                            timeout_seconds: Some(invocation.timeout.as_secs()),
-                            budget_exhausted: None,
-                        };
-
-                        let mut warnings = Vec::new();
-                        match llm_err {
-                            crate::llm::LlmError::Timeout { duration } => {
-                                llm_info.timed_out = Some(true);
-                                llm_info.timeout_seconds = Some(duration.as_secs());
-                                warnings.push(format!("phase_timeout:{}", duration.as_secs()));
-                            }
-                            _ => {
-                                llm_info.timed_out = Some(false);
-                                warnings.push(format!("llm_error:{}", llm_err));
-                            }
-                        }
-                        if let Some(ref warning) = llm_fallback_warning {
-                            warnings.push(warning.clone());
-                        }
-
-                        let mut receipt = self.receipt_manager().create_receipt_with_redactor(
-                            config.redactor.as_ref(),
-                            self.spec_id(),
-                            phase_id,
-                            exit_code,
-                            vec![], // No successful outputs
-                            env!("CARGO_PKG_VERSION"),
-                            "unknown", // Claude hasn't completed, so version is unknown
-                            configured_model,
-                            None, // No model alias
-                            flags,
-                            packet_evidence,
-                            None, // No stderr_tail
-                            None, // No stderr_redacted
-                            warnings,
-                            None, // No fallback
-                            configured_runner,
-                            None, // No runner distro
-                            Some(error_kind),
-                            Some(llm_err.to_string()),
-                            None, // No diff_context
-                            pipeline_info.clone(),
-                        );
-
-                        receipt.llm = Some(llm_info);
-
-                        let receipt_path = self.receipt_manager().write_receipt(&receipt)?;
-
-                        return Ok(ExecutionResult {
-                            phase: phase_id,
-                            success: false,
-                            exit_code,
-                            artifact_paths: vec![],
-                            receipt_path: Some(receipt_path.into_std_path_buf()),
-                            error: Some(llm_err.to_string()),
-                        });
-                    }
-                    // For other errors, propagate normally
-                    return Err(e);
-                }
-            }
-        };
+                PhaseLlmOutcome::Completed {
+                    claude_response,
+                    claude_exit_code,
+                    claude_metadata,
+                    llm_result,
+                    llm_fallback_warning,
+                } => (
+                    claude_response,
+                    claude_exit_code,
+                    claude_metadata,
+                    llm_result,
+                    llm_fallback_warning,
+                ),
+                PhaseLlmOutcome::Failed(result) => return Ok(result),
+            };
 
         // Handle Claude CLI failure (R4.3)
         if claude_exit_code != 0 {
             // Save partial output as required by R4.3
-            let partial_filename = format!(
-                "{:02}-{}.partial.md",
-                self.get_phase_number(phase_id),
-                phase_id.as_str().to_lowercase()
-            );
-
-            let partial_result = self.artifact_manager().store_artifact(&Artifact {
-                name: partial_filename.clone(),
-                content: claude_response.clone(),
-                artifact_type: ArtifactType::Partial,
-                blake3_hash: blake3::hash(claude_response.as_bytes())
-                    .to_hex()
-                    .to_string(),
-            })?;
-            let partial_path = partial_result.path;
+            let (partial_filename, partial_path) =
+                self.store_failed_llm_partial(phase_id, &claude_response)?;
 
             // Create failure receipt with stderr_tail and warnings (R4.3)
             // Use the actual packet evidence from the packet that was created
@@ -1226,7 +795,7 @@ impl PhaseOrchestrator {
                 phase: phase_id,
                 success: false,
                 exit_code: claude_exit_code,
-                artifact_paths: vec![partial_path.into_std_path_buf()], // Include partial artifact
+                artifact_paths: vec![partial_path], // Include partial artifact
                 receipt_path: Some(receipt_path.into_std_path_buf()),
                 error: Some(enhanced_error.to_string()),
             });
@@ -1242,65 +811,8 @@ impl PhaseOrchestrator {
                 )
             })?;
 
-        // Step 7: Write partial artifacts to .partial/ subdirectory (FR-ORC-004)
-        let mut artifact_paths = Vec::new();
-        let mut output_hashes = Vec::new();
-        let mut atomic_write_warnings = Vec::new();
-
-        for artifact in &phase_result.artifacts {
-            // Store to .partial/ staging directory first
-            let partial_result = self
-                .artifact_manager()
-                .store_partial_staged_artifact(artifact)
-                .with_context(|| format!("Failed to store partial artifact: {}", artifact.name))?;
-
-            // Collect atomic write warnings
-            for warning in &partial_result.atomic_write_result.warnings {
-                atomic_write_warnings.push(format!("{}: {}", artifact.name, warning));
-            }
-
-            // Create file hash for receipt (using final content)
-            // Determine file type from extension for proper canonicalization
-            let file_type = if let Some(ext) = std::path::Path::new(&artifact.name).extension() {
-                FileType::from_extension(ext.to_str().unwrap_or(""))
-            } else {
-                // Fallback to artifact type if no extension
-                match artifact.artifact_type {
-                    ArtifactType::Markdown => FileType::Markdown,
-                    ArtifactType::CoreYaml => FileType::Yaml,
-                    _ => FileType::Text,
-                }
-            };
-
-            let file_hash = self
-                .receipt_manager()
-                .create_file_hash(
-                    &format!("artifacts/{}", artifact.name),
-                    &artifact.content,
-                    file_type,
-                    phase_id.as_str(),
-                )
-                .map_err(|e| {
-                    XCheckerError::Phase(PhaseError::OutputValidationFailed {
-                        phase: phase_id.as_str().to_string(),
-                        reason: e.to_string(),
-                    })
-                })?;
-
-            output_hashes.push(file_hash);
-        }
-
-        // Step 8: Promote to final (atomic rename) (FR-ORC-004)
-        for artifact in &phase_result.artifacts {
-            let final_path = self
-                .artifact_manager()
-                .promote_staged_to_final(&artifact.name)
-                .with_context(|| {
-                    format!("Failed to promote artifact to final: {}", artifact.name)
-                })?;
-
-            artifact_paths.push(final_path.into_std_path_buf());
-        }
+        // Steps 7-8: Stage, hash, and promote successful artifacts (FR-ORC-004).
+        let stored_artifacts = self.store_success_artifacts(phase_id, &phase_result.artifacts)?;
 
         // Step 9: Create and write receipt (FR-ORC-005, FR-ORC-006)
         // Use the actual packet evidence from the packet that was created
@@ -1318,7 +830,8 @@ impl PhaseOrchestrator {
             (None, "haiku".to_string())
         };
 
-        let mut warnings: Vec<String> = atomic_write_warnings
+        let mut warnings: Vec<String> = stored_artifacts
+            .atomic_write_warnings
             .into_iter()
             .chain(hook_warnings.iter().cloned())
             .collect();
@@ -1331,7 +844,7 @@ impl PhaseOrchestrator {
             self.spec_id(),
             phase_id,
             0, // Success exit code
-            output_hashes,
+            stored_artifacts.output_hashes,
             env!("CARGO_PKG_VERSION"),
             claude_metadata
                 .as_ref()
@@ -1363,62 +876,14 @@ impl PhaseOrchestrator {
             .write_receipt(&receipt)
             .with_context(|| format!("Failed to write receipt for phase: {}", phase_id.as_str()))?;
 
-        // Execute post-phase hook if configured (runs on success)
-        // Hooks run from invocation CWD so relative paths like ./scripts/... work
-        // Note: Post-hook failures are treated as warnings, not phase failures
-        // (artifacts have already been created and receipt written)
-        if let Some(ref hooks_config) = config.hooks
-            && let Some(hook_config) = hooks_config.get_post_phase_hook(phase_id)
-        {
-            let executor = HookExecutor::new(
-                std::env::current_dir().unwrap_or_else(|_| std::path::PathBuf::from(".")),
-            );
-            let context = HookContext::new(self.spec_id(), phase_id, HookType::PostPhase);
-
-            match execute_and_process_hook(
-                &executor,
-                hook_config,
-                &context,
-                HookType::PostPhase,
-                phase_id,
-            )
-            .await
-            {
-                Ok(outcome) => {
-                    // Log any warnings from successful or failed hooks
-                    if let Some(warning) = outcome.warning() {
-                        tracing::warn!(
-                            phase = %phase_id.as_str(),
-                            "Post-phase hook warning: {}",
-                            warning.to_warning_string()
-                        );
-                    }
-                    // Check if hook wanted to fail but we treat it as warning
-                    // (post-hooks run after artifacts are created, so we don't fail the phase)
-                    if !outcome.should_continue() {
-                        tracing::warn!(
-                            phase = %phase_id.as_str(),
-                            "Post-phase hook had on_fail=fail but phase artifacts already created; treating as warning"
-                        );
-                    }
-                }
-                Err(e) => {
-                    // Log hook execution errors but don't fail the phase
-                    // (artifacts are already created at this point)
-                    tracing::warn!(
-                        phase = %phase_id.as_str(),
-                        error = %e,
-                        "Post-phase hook execution error (treated as warning)"
-                    );
-                }
-            }
-        }
+        // Execute post-phase hook after artifacts and receipt are written.
+        self.run_post_phase_hook(phase_id, config).await;
 
         Ok(ExecutionResult {
             phase: phase_id,
             success: true,
             exit_code: 0,
-            artifact_paths,
+            artifact_paths: stored_artifacts.artifact_paths,
             receipt_path: Some(receipt_path.into_std_path_buf()),
             error: None,
         })
